Change group id
Change dependencies
