+++ src/main/kotlin/io/mahdibohloul/kpringmediator/infrastructure/RequestHandlerProvider.kt
@@ -1,17 +1,17 @@
-package io.kpring.mediator.infrastructure
+package io.mahdibohloul.kpringmediator.infrastructure
 
-import io.kpring.mediator.core.CommandHandler
+import io.mahdibohloul.kpringmediator.core.RequestHandler
 import org.springframework.context.ApplicationContext
 import kotlin.reflect.KClass
 
 /**
- * A wrapper around [CommandHandler]
+ * A wrapper around [RequestHandler]
  *
  * @author Mahdi Bohloul
  * @property applicationContext ApplicationContext from Spring used to retrieve beans
  * @property type Type of CommandHandler
  */
-internal class CommandHandlerProvider<T : CommandHandler<*>>(
+class RequestHandlerProvider<T : RequestHandler<*, *>>(
     private val applicationContext: ApplicationContext,
     private val type: KClass<T>
 ) {
