+++ src/test/kotlin/io/mahdibohloul/kpringmediator/mock/CommandMock.kt
@@ -1,9 +1,7 @@
-package io.kpring.mediator.mock
+package io.mahdibohloul.kpringmediator.mock
 
-import io.kpring.mediator.core.Command
-import io.kpring.mediator.core.CommandHandler
-import io.kpring.mediator.core.Notification
-import io.kpring.mediator.core.NotificationHandler
+import io.mahdibohloul.kpringmediator.core.Command
+import io.mahdibohloul.kpringmediator.core.CommandHandler
 import org.slf4j.LoggerFactory
 import org.springframework.stereotype.Component
